requirements.txt
simple-term-menu == 1.6.3
pika             == 1.3.2
grpcio           == 1.59.2
protobuf         == 4.25.0